Add test coverage of missing schema, fix error cb
Adds a specific test for handling firehose events tagged with a stream
name for which we don't have an avro schema. Asserts that the invocation
calls back with an error and logs helpful info.

This work uncovered that several thrown errors were sending the error
message as the error callback, which is ~fine~ but not quite correct;
Changed them to pass the whole error back. This required updating
several tests of error conditions.